Edit code style
