+++ practice/practise-4/build.gradle.kts
@@ -14,6 +14,7 @@ checkstyle {
 	isIgnoreFailures = false
 	maxWarnings = 0
 	maxErrors = 0
+	configFile = file("checkstyle/checkstyle.xml")
 }
 
 java {
