+++ practice/practise-4/src/main/java/hse/kpo/domains/Car.java
@@ -26,9 +26,13 @@ public class Car {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj) return true; // Сравнение по ссылке
+        if (this == obj) {
+            return true; // Сравнение по ссылке
+        }
 
-        if (obj == null || getClass() != obj.getClass()) return false; // Проверка типа
+        if (obj == null || getClass() != obj.getClass()) {
+            return false; // Проверка типа
+        }
 
         var car = (Car) obj;
         return vin == car.vin && engine.getClass() == car.engine.getClass(); // Сравнение полей
