+++ practice/practise-4/src/main/java/hse/kpo/domains/Customer.java
@@ -1,10 +1,10 @@
 package hse.kpo.domains;
 
+import java.util.Objects;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-import java.util.Objects;
 
 /**
  * Customer class.
@@ -35,12 +35,17 @@ public class Customer {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj) return true; // Сравнение по ссылке
+        if (this == obj) {
+            return true;         // Сравнение по ссылке
+        }
 
-        if (obj == null || getClass() != obj.getClass()) return false; // Проверка типа
+
+        if (obj == null || getClass() != obj.getClass()) {
+            return false;
+        } // Проверка типа
 
         var customer = (Customer) obj;
-        return Objects.equals(name, customer.getName()) && legPower == customer.getLegPower() &&
-                handPower == customer.getHandPower() && iq == customer.getIq(); // Сравнение полей
+        return Objects.equals(name, customer.getName()) && legPower == customer.getLegPower()
+                && handPower == customer.getHandPower() && iq == customer.getIq(); // Сравнение полей
     }
 }
