--- practice/practise-4/src/main/java/hse/kpo/interfaces/ICarFactory.java
@@ -1,7 +1,0 @@
-package hse.kpo.interfaces;
-
-import hse.kpo.domains.Car;
-
-public interface ICarFactory<TParams> {
-    Car createCar(TParams carParams, int carNumber);
-}
