--- practice/practise-4/src/main/java/hse/kpo/interfaces/ICarProvider.java
@@ -1,9 +1,0 @@
-package hse.kpo.interfaces;
-
-import hse.kpo.domains.Car;
-import hse.kpo.domains.Customer;
-
-public interface ICarProvider {
-
-    Car takeCar(Customer customer); // Метод возвращает optional на Car, что означает, что метод может ничего не вернуть
-}
