--- practice/practise-4/src/main/java/hse/kpo/interfaces/IEngine.java
@@ -1,15 +1,0 @@
-package hse.kpo.interfaces;
-
-
-import hse.kpo.domains.Customer;
-
-public interface IEngine {
-
-    /**
-     * Метод для проверки совместимости двигателя с покупателем.
-     *
-     * @param customer - покупатель, с которым мы сравниваем двигатель
-     * @return true, если двигатель подходит покупателю
-     */
-    boolean isCompatible(Customer customer);
-}
