+++ practice/practise-4/src/main/java/hse/kpo/services/CarService.java
@@ -6,7 +6,6 @@ import hse.kpo.interfaces.InterfaceCarFactory;
 import hse.kpo.interfaces.InterfaceCarProvider;
 import java.util.ArrayList;
 import java.util.List;
-
 import lombok.Getter;
 import org.springframework.stereotype.Component;
